factored out WheelControllerBase and GeomControllerBase
